Struct kubecfg::Cluster

pub struct Cluster {
    pub api_version: String,
    pub server: String,
    pub insecure_skip_tls_verify: bool,
    pub certificate_authority: Option<Content>,
}

Describes information needed to resolve a connection to a cluster

Fields

api_version

The clusters supported api version

server

The server URI

insecure_skip_tls_verify

Predicate used to determine if a client should skip tls verification

certificate_authority

Content used by client to certify the server is authentic

Methods

impl Cluster

fn new(api_version: Option<String>, server: Option<String>, insecure_skip_tls_verify: Option<bool>, certificate_authority: Option<Content>) -> Cluster