Struct kubecfg::Config

pub struct Config {
    pub clusters: HashMap<String, Cluster>,
    pub contexts: HashMap<String, Context>,
    pub users: HashMap<String, User>,
    pub current_context: String,
}

Represents local kubernetes configuration settings

Fields

clusters

A map of cluster name to cluster

contexts

A map of context name to context

users

A map of user name to user

current_context

The current context's name

Methods

impl Config

fn from_std_path() -> Result<Config>

Reads a Config object from the default location on disk

fn from_path<P>(path: P) -> Result<Config> where P: AsRef<Path>

Reads a Config object from a custom location on disk

fn from_str(raw: &str) -> Result<Config>

Reads a Config object from a raw string payload